Move function info into separate file.
